feat(core): Split unix main, link core obj into Rust test target
