+++ core/embed/rust/build.rs
@@ -1,8 +1,12 @@
+#[cfg(feature = "test")]
+use std::ffi::OsStr;
 use std::{env, path::PathBuf, process::Command};
 
 fn main() {
     generate_qstr_bindings();
     generate_micropython_bindings();
+    #[cfg(feature = "test")]
+    link_core_objects();
 }
 
 /// Generates Rust module that exports QSTR constants used in firmware.
@@ -163,6 +167,9 @@ fn generate_micropython_bindings() {
         .use_core()
         .ctypes_prefix("cty")
         .size_t_is_usize(true)
+        // Disable the layout tests. They spew out a lot of code-style bindings, and are not too
+        // relevant for our use-case.
+        .layout_tests(false)
         // Tell cargo to invalidate the built crate whenever any of the
         // included header files change.
         .parse_callbacks(Box::new(bindgen::CargoCallbacks))
@@ -176,4 +183,29 @@ fn generate_micropython_bindings() {
 fn is_firmware() -> bool {
     let target = env::var("TARGET").unwrap();
     target.starts_with("thumbv7")
+}
+
+#[cfg(feature = "test")]
+fn link_core_objects() {
+    let crate_path = env::var("CARGO_MANIFEST_DIR").unwrap();
+    let build_path = format!("{}/../../build/unix", crate_path);
+
+    let embed_blocklist = [OsStr::new("main_main.o")];
+
+    let mut cc = cc::Build::new();
+    for obj in glob::glob(&format!("{}/embed/**/*.o", build_path)).unwrap() {
+        let obj = obj.unwrap();
+        if embed_blocklist.contains(&obj.file_name().unwrap()) {
+            // Ignore.
+        } else {
+            cc.object(obj);
+        }
+    }
+    for obj in glob::glob(&format!("{}/vendor/**/*.o", build_path)).unwrap() {
+        cc.object(obj.unwrap());
+    }
+    cc.compile("core_lib");
+
+    println!("cargo:rustc-link-lib=SDL2");
+    println!("cargo:rustc-link-lib=SDL2_image");
 }
